# openapi.json
{
    "openapi": "3.0.3",
    "info": {
        "title": "Pharmacy Management System API",
        "version": "1.0.0",
        "description": "Multi-tenant Pharmacy Management System"
    },
    "paths": {
...
        "/api/logout/": {
            "post": {
                "operationId": "logout_create",
                "tags": [
                    "logout"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
        "/api/owner/create-user/": {
            "post": {
                "operationId": "owner_create_user_create",
                "tags": [
                    "owner"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
...
        "/api/owner/notifications/": {
            "get": {
                "operationId": "owner_notifications_list",
                "description": "List notifications for the tenant. Use `filter=read|unread` to filter.",
                "tags": [
                    "owner"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/NotificationModel"
                                    }
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "post": {
                "operationId": "owner_notifications_create",
                "description": "Create a notification for the tenant (recipient optional).",
                "tags": [
                    "owner"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationModelRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationModelRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/NotificationModelRequest"
                            }
                        }
                    },
                    "required": true
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationModel"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            }
        },
        "/api/owner/notifications/{notification_id}/": {
            "patch": {
                "operationId": "owner_notifications_partial_update",
                "parameters": [
                    {
                        "in": "path",
                        "name": "notification_id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "owner"
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedNotificationMarkRequest"
                            }
                        },
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedNotificationMarkRequest"
                            }
                        },
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/PatchedNotificationMarkRequest"
                            }
                        }
                    }
                },
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotificationModel"
                                }
                            }
                        },
                        "description": ""
                    }
                }
            },
            "delete": {
                "operationId": "owner_notifications_destroy",
                "description": "Delete a notification",
                "parameters": [
                    {
                        "in": "path",
                        "name": "notification_id",
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        },
                        "required": true
                    }
                ],
                "tags": [
                    "owner"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No response body"
                    }
                }
            }
        },
...
        "/api/owner/users/": {
            "get": {
                "operationId": "owner_users_retrieve",
                "tags": [
                    "owner"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
...
        "/api/users/search/": {
            "get": {
                "operationId": "users_search_retrieve",
                "tags": [
                    "users"
                ],
                "security": [
                    {
                        "jwtAuth": []
                    },
                    {
                        "cookieAuth": []
                    },
                    {
                        "tokenAuth": []
                    }
                ],
                "responses": {
                    "200": {
                        "description": "No response body"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NotificationModel": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "tenantId": {
                        "type": "string",
                        "format": "uuid",
                        "readOnly": true
                    },
                    "title": {
                        "type": "string",
                        "maxLength": 255
                    },
                    "message": {
                        "type": "string"
                    },
                    "recipientId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "isRead": {
                        "type": "boolean"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "required": [
                    "createdAt",
                    "id",
                    "isRead",
                    "message",
                    "tenantId",
                    "title"
                ]
            },
            "NotificationModelRequest": {
                "type": "object",
                "properties": {
                    "title": {
                        "type": "string",
                        "minLength": 1,
                        "maxLength": 255
                    },
                    "message": {
                        "type": "string",
                        "minLength": 1
                    },
                    "recipientId": {
                        "type": "string",
                        "format": "uuid",
                        "nullable": true
                    },
                    "isRead": {
                        "type": "boolean"
                    }
                },
                "required": [
                    "isRead",
                    "message",
                    "title"
                ]
            },
...
            "PatchedNotificationMarkRequest": {
                "type": "object",
                "properties": {
                    "isRead": {
                        "type": "boolean"
                    }
                }
            },
...
        },
        "securitySchemes": {
            "cookieAuth": {
                "type": "apiKey",
                "in": "cookie",
                "name": "sessionid"
            },
            "jwtAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            },
            "tokenAuth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Token-based authentication with required prefix \"Token\""
            }
        }
    }
}